Change svg entity model

+++ PcbReader/src/Layers/Gerber/Entities/Apertures/CircleAperture.cs
@@ -1,6 +1,6 @@
 ﻿namespace PcbReader.Layers.Gerber.Entities.Apertures;
 
 public class CircleAperture : IAperture {
-    public decimal Diameter { get; set; }
-    public decimal? HoleDiameter { get; set; }
+    public double Diameter { get; set; }
+    public double? HoleDiameter { get; set; }
 }
