+++ PcbReader/src/Layers/Gerber/Entities/Apertures/ObRoundAperture.cs
@@ -1,7 +1,7 @@
 ﻿namespace PcbReader.Layers.Gerber.Entities.Apertures;
 
 public class ObRoundAperture: IAperture {
-    public decimal XSize { get; set; }
-    public decimal YSize { get; set; }
-    public decimal? HoleDiameter { get; set; }
+    public double XSize { get; set; }
+    public double YSize { get; set; }
+    public double? HoleDiameter { get; set; }
 }
