+++ PcbReader/src/Layers/Gerber/Entities/Apertures/PolygonAperture.cs
@@ -1,8 +1,8 @@
 ﻿namespace PcbReader.Layers.Gerber.Entities.Apertures;
 
 public class PolygonAperture: IAperture {
-    public decimal OuterDiameter { get; set; }
+    public double OuterDiameter { get; set; }
     public int VerticesCount { get; set; }
-    public decimal? Rotation { get; set; }
-    public decimal? HoleDiameter { get; set; }
+    public double? Rotation { get; set; }
+    public double? HoleDiameter { get; set; }
 }
